+++ projects/fhi-angular-highcharts/src/lib/fhi-diagram-serie-name-seperator.constant.ts
@@ -1,0 +1,4 @@
+export class FhiDiagramSerieNameSeperator {
+  static in = '|';
+  static out = ' | ';
+};