+++ projects/fhi-angular-highcharts/src/lib/services/options.service.ts
@@ -5,13 +5,13 @@ import merge from 'lodash-es/merge';
 import { Options, SeriesOptionsType, XAxisLabelsOptions, XAxisOptions } from 'highcharts';
 import { isValid, parseISO } from 'date-fns'
 
-import { FhiAllDiagramTypes } from '../fhi-diagram/fhi-diagram-type.constants';
+import { FhiAllDiagramTypes } from '../fhi-diagram-type.constants';
 // import { GeoJsonService } from './geo-json.service';
-import { FhiDiagramOptions, FhiDiagramSerie } from '../fhi-diagram/fhi-diagram.models';
+import { FhiDiagramOptions, FhiDiagramSerie } from '../fhi-diagram.models';
 import { OptionsChartsAndMaps } from '../highcharts-options/options-charts-and-maps';
 import { OptionsCharts } from '../highcharts-options/options-charts';
 import { OptionsMaps } from '../highcharts-options/options-maps';
-import { FhiDiagramType } from '../fhi-diagram/fhi-diagram.models'
+import { FhiDiagramType } from '../fhi-diagram.models'
 
 @Injectable({
   providedIn: 'root'
@@ -53,13 +53,13 @@ export class OptionsService {
     return merge(chartsAndMaps, current, optionsCurrentChartType);
   }
 
-  private getSeries(data: FhiDiagramSerie[], isMap: boolean | undefined, allMapsLoaded: boolean): SeriesOptionsType[] {
+  private getSeries(series: FhiDiagramSerie[], isMap: boolean | undefined, allMapsLoaded: boolean): SeriesOptionsType[] {
     if (isMap && allMapsLoaded) {
       // TODO: MAP
       // return [this.geoJsonService.getHighmapsSerie(series[0])];
 
     } else {
-      const highchartsSeries = cloneDeep(data);
+      const highchartsSeries = cloneDeep(series);
       highchartsSeries.forEach((serie) => {
         // Remove flagged data from Highcharts options series
         serie.data = serie.data.filter(dataPoint =>  typeof dataPoint.y !== 'string')