+++ projects/fhi-angular-highcharts/CHANGELOG.md
@@ -1,4 +1,12 @@
-# 0.5.0
+# Unreleased
+
+> Jun 21, 2023
+
+* :tada: **Enhancement** Change seperator in `serie.name` to `|`
+* :tada: **Enhancement** Add support for type `Array<string>` in `serie.name`
+* :tada: **Enhancement** Update documentation
+
+## 0.5.0
 
 > Jun 15, 2023